% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_group_to_pyproject.R
\name{append_dep_groups_table}
\alias{append_dep_groups_table}
\title{Append a fresh \verb{[dependency-groups]} table at end of file}
\usage{
append_dep_groups_table(lines, name, deps)
}
\description{
Append a fresh \verb{[dependency-groups]} table at end of file
}
\keyword{internal}
