% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_group_to_pyproject.R
\name{append_subgroup}
\alias{append_subgroup}
\title{Append a subgroup inside an existing \verb{[dependency-groups]} table}
\usage{
append_subgroup(lines, groups, name, deps)
}
\description{
Append a subgroup inside an existing \verb{[dependency-groups]} table
}
\keyword{internal}
