% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_group_to_pyproject.R
\name{pkg_name_of}
\alias{pkg_name_of}
\title{Extract the package name from a PEP 508-ish pin string}
\usage{
pkg_name_of(deps)
}
\description{
Handles \code{pkg}, \code{pkg==X.Y.Z}, \code{pkg>=X}, \code{pkg[extra]==X}, etc. Strips
anything after the first occurrence of \code{[}, \code{=}, \code{<}, \code{>}, \code{!}, \code{~},
\verb{;}, or whitespace.
}
\keyword{internal}
