% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_python.R
\name{read_requires_python}
\alias{read_requires_python}
\title{Read \code{requires-python} from \verb{[project]} table}
\usage{
read_requires_python(lines)
}
\description{
Read \code{requires-python} from \verb{[project]} table
}
\keyword{internal}
