% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_group_to_pyproject.R
\name{render_subgroup}
\alias{render_subgroup}
\title{Render a \verb{[dependency-groups]} subgroup in canonical multi-line form}
\usage{
render_subgroup(name, deps)
}
\description{
Render a \verb{[dependency-groups]} subgroup in canonical multi-line form
}
\keyword{internal}
